Analysis update: account for or remove interannual variation
When digging into the model performing lower than expected and talking with Ross, I realized that it's ESSENTIAL to account for interannual variation since ET can be highly influenced by interannual variation in moisture availability etc.  Previously, I'd avoided adding a year intercept because then you have to specify a year to predict without making the code for prediction a whole lot more complicated.  Because of those potential complications, I'm also running and saving a model with the average temp, cover, and ET, but it'd be nice not to have to use it.