First stab at looking at summaries
all of the warming projections are for LOWER ET... that's odd... I think it's because of lower ET in warmer parts of the city.  This can come from covariance between LST and low tree cover, but also with trees tending to shut down at higher temps... We need to interrogate this more!  New script coming!