Summary analyses!
Looked at the partial effects & impacts of temperature in the cmip6 scenarios -- then went back and rejiggered to predict the ET holding other values at their mean.  Because the tree effect is so big and varies so much more than temp, it really makes the absolute effect of temperature range look small.  But the effect of temperature will be bigger at high canopy conditions