Add script to extract IGBP cover for cities
I think that the tower IGBP designations don't match the characteristics of the broader area I'm trying to model.  So I've pulled %cover from a MODIS product (which I think partially informs the ET model?)  for both the pixel of an individual tower and the city as a whole to characterize the mismatch.  I decided to go this route over looking at the tree cover etc. of the pixel.